Implement ControlValueAccessor directly with the component
Instead of using a separate class for implementing the ControlValueAccessor interface
and injecting it into the component, let the component implement it.

This appears to work, but seems to cause another bug:
Updating properties on the ngModel-bound object by typing in any of the input elements
causes the HTML to rerender, blurring the input fields. Might also be caused by
using the [ngFor] index to bind the property <input>s:

    <div *ngFor="customer of customers; #n = index">
        <input [(ngModel)]="customers[n].name">
    </div>